+++ src/utility/nodebuilder/nodebuild_utility.cpp
@@ -460,7 +460,7 @@ void FNodeBuilder::FindPolyContainers (TArray<FPolyStart> &spots, TArray<FPolySt
 						if (dist < closestdist && dist >= 0)
 						{
 							closestdist = dist;
-							closestseg = (long)j;
+							closestseg = j;
 						}
 					}
 				}
